split Listener to TCP and UDP Listener
